Renamed config package to internal in the ZWay binding. See openhab/static-code-analysis#100
