+++ addons/binding/org.openhab.binding.zway/src/main/java/org/openhab/binding/zway/handler/ZWayBridgeHandler.java
@@ -26,7 +26,7 @@ import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.binding.BaseBridgeHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.types.Command;
-import org.openhab.binding.zway.config.ZWayBridgeConfiguration;
+import org.openhab.binding.zway.internal.config.ZWayBridgeConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
