+++ addons/binding/org.openhab.binding.zway/src/main/java/org/openhab/binding/zway/handler/ZWayZAutomationDeviceHandler.java
@@ -21,7 +21,7 @@ import org.eclipse.smarthome.core.thing.ThingStatus;
 import org.eclipse.smarthome.core.thing.ThingStatusDetail;
 import org.eclipse.smarthome.core.thing.ThingStatusInfo;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
-import org.openhab.binding.zway.config.ZWayZAutomationDeviceConfiguration;
+import org.openhab.binding.zway.internal.config.ZWayZAutomationDeviceConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
