+++ addons/binding/org.openhab.binding.zway/src/main/java/org/openhab/binding/zway/handler/ZWayZWaveDeviceHandler.java
@@ -22,7 +22,7 @@ import org.eclipse.smarthome.core.thing.ThingStatus;
 import org.eclipse.smarthome.core.thing.ThingStatusDetail;
 import org.eclipse.smarthome.core.thing.ThingStatusInfo;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
-import org.openhab.binding.zway.config.ZWayZWaveDeviceConfiguration;
+import org.openhab.binding.zway.internal.config.ZWayZWaveDeviceConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
